+++ src/main/java/net/sf/jabref/gui/entryeditor/FieldExtraComponents.java
@@ -46,7 +46,6 @@ import net.sf.jabref.gui.entryeditor.EntryEditor.StoreFieldAction;
 import net.sf.jabref.gui.fieldeditors.FieldEditor;
 import net.sf.jabref.gui.mergeentries.MergeEntryDOIDialog;
 import net.sf.jabref.gui.undo.UndoableFieldChange;
-import net.sf.jabref.importer.fetcher.CrossRef;
 import net.sf.jabref.logic.journals.JournalAbbreviationRepository;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.net.URLUtil;
@@ -187,12 +186,12 @@ public class FieldExtraComponents {
         // lookup doi
         JButton doiButton = new JButton(Localization.lang("Lookup DOI"));
         doiButton.addActionListener(actionEvent -> {
-                Optional<DOI> doi = CrossRef.findDOI(entryEditor.getEntry());
-                if (doi.isPresent()) {
-                    entryEditor.getEntry().setField("doi", doi.get().getDOI());
-                } else {
-                    panel.frame().setStatus(Localization.lang("No DOI found"));
-                }
+            Optional<DOI> doi = DOI.fromBibEntry(entryEditor.getEntry());
+            if (doi.isPresent()) {
+                entryEditor.getEntry().setField("doi", doi.get().getDOI());
+            } else {
+                panel.frame().setStatus(Localization.lang("No DOI found"));
+            }
         });
         // fetch bibtex data
         JButton fetchButton = new JButton(Localization.lang("Get BibTeX data from DOI"));